Complete Spotify Authentication
First, I edited the migration for the users table so there are places
for the access and refresh tokens.

Then, I added a route for the callback part of the authentication
process.

Finally, I added a callback function in the API connection controller
class that saves the the access and refresh tokens in the database if
that user doesn't already exist in the database.