Update many files because of database changes
One of the things that I did in the new migrations is make the names of
some fields more specific.

uid => user_id
sid => song_id
tid => trend_id
songs.name => songs.song_name
users.name => users.name (unchanged)

The first three are trivial, but I like that they are slightly more
differentiated now. Changing the name field in the songs table removes
ambiguity that I encountered in the code where it is kind of hard to
tell which name is being referenced.